Adding an indicator in QR_matrix objects

Function to add indicators in QR_matrix objects.

add_indicator(x, indicator, variable_name, ...)

Arguments

x

a QR_matrix or mQR_matrix object

indicator

a vector or a data.frame (cf. details).

variable_name

a string containing the name of the variables to add.

...

other parameters of the function merge.

Details

The function add_indicator() adds a chosen indicator in the values matrix of a quality report. Therefore, because said indicator isn't added in the modalities matrix, it cannot be used to calculate a score (except for weighting). Before using the added variable for score calculation, the function recode_indicator_num will have to be adapted.

The new indicator can be a vector or a data.frame. In both cases, its format must allow for pairing:

a vector's elements must be named and these names must match those of the quality report (variable "series");

a data.frame must contain a "series" column that matches with the quality report's series.

See also