Creation of the .params file for the cruncher

To operate, the cruncher needs a parameter file. It can be generated by the function create_param_file.

create_param_file(
  dir_file_param = getwd(),
  bundle = 10000,
  csv_layout = "list",
  csv_separator = ";",
  ndecs = 6,
  policy = "parameters",
  output = NULL,
  matrix_item = getOption("default_matrix_item"),
  tsmatrix_series = getOption("default_tsmatrix_series"),
  paths_path = NULL
)

Arguments

dir_file_param

path to the folder that will store the .param parameter file.

bundle

maximum number of series in an output file. By default, bundle = 10000.

csv_layout

layout of the output file. By default, csv_layout = "list". Other options are: csv_layout = "vtable" or csv_layout = "htable".

csv_separator

column separator for the csv file. By default, csv_separator = ";".

ndecs

number of decimals in the output (6 by default).

policy

refresh policy used. By default, policy = "parameters" (the model coefficients are re-estimated). The other methods available are: "current" or "n" (fixed model + all new data are classified as additive outliers) "fixed" or "f" (fixed model: the model is stictly unchanged) "fixedparameters" or "fp" (re-estimation of the regression coefficients) "parameters" or "p" (above + re-estimation of the arima coefficients) "lastoutliers" or "l" (above + re-identification of the outliers over the last year) ; "outliers" or "o" (above + re-identification of the outliers over the whole series span) ; "stochastic" or "s" (above + re-estimation of the arima model (orders)) ; "complete" or "c" (the model is completely re-estimated).

output

folder where the results are stored. By default, (output = NULL) and a new "Output" folder is created in the workspace folder.

matrix_item

string speciying the parameters to export (cf. JDemetra+ documentation). To get the default parameters, compile getOption("default_matrix_item"). This option is initially set to JDemetra+'s default value.

tsmatrix_series

string specifying the time series to export (cf. JDemetra+ documentation). To get the default parameters, compile getOption("default_tsmatrix_series"). This option is initially set to JDemetra+'s default value.

paths_path

paths to the input files (Excel, xml...).

Value

The parameters file address.

See also