Extraction d'un bilan qualité

Permet d'extraire un bilan qualité à partir du fichier CSV contenant la matrice des diagnostics.

Arguments

matrix_output_file

Chaîne de caracère. Chemin vers le fichier CSV contenant la matrice des diagnostics.

file

Chaîne de caracère. Chemin vers le fichier CSV contenant la matrice des diagnostics. Cet argument remplace l'argument matrix_output_file.

sep

séparateur de caractères utilisé dans le fichier csv (par défaut sep = ";")

dec

séparateur décimal utilisé dans le fichier csv (par défaut dec = ",")

thresholds

list de vecteurs numériques. Seuils appliqués aux différents tests afin de classer en modalités Good, Uncertain, Bad et Severe. Par défault, la valeur de l'option "jdc_threshold" est utilisée. Vous pouvez appeler la fonction get_thresholds pour voir à quoi doit ressemble l'objet thresholds.

Value

Un objet de type QR_matrix.

Details

La fonction permet d'extraire un bilan qualité à partir d'un fichier csv contenant l'ensemble des diagnostics (généralement fichier demetra_m.csv).

Ce fichier peut être obtenu en lançant le cruncher (cruncher ou cruncher_and_param) avec l'ensemble des paramètres de base pour les paramètres à exporter et l'option csv_layout = "vtable" (par défaut) pour le format de sortie des fichiers csv (option de cruncher_and_param ou de create_param_file lors de la création du fichier de paramètres).

Le résultat de cette fonction est un objet QR_matrix qui est une liste de trois paramètres :

le paramètre modalities est un data.frame contenant un ensemble de variables sous forme catégorielle (Good, Uncertain, Bad, Severe).

le paramètre values est un data.frame contenant les valeurs associées aux indicateurs présents dans modalities (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas de modalité (fréquence de la série et modèle ARIMA).

le paramètre score_formula est initié à NULL : il contiendra la formule utilisée pour calculer le score (si le calcul est fait).

Si x est fourni, les arguments fichier et matrix_output_file sont ignorés. L'argument fichier désigne également le chemin vers le fichier qui contient la matrice de diagnostic (qui peut être importée en parallèle dans R et utilisée avec l'argument x).

Examples

# Chemin menant au fichier demetra_m.csv
demetra_path <- file.path(
    system.file("extdata", package = "JDCruncheR"),
    "WS/ws_ipi/Output/SAProcessing-1",
    "demetra_m.csv"
)

# Extraire le bilan qualité à partir du fichier demetra_m.csv
QR <- extract_QR(file = demetra_path)

print(QR)
#> The quality report matrix has 13 observations
#> There are 18 indicators in the modalities matrix and 20 indicators in the values matrix
#> 
#> The quality report matrix contains the following variables:
#> series  residuals_homoskedasticity  residuals_skewness  residuals_kurtosis  qs_residual_sa_on_sa  f_residual_sa_on_sa  qs_residual_sa_on_i  f_residual_sa_on_i  f_residual_td_on_sa  f_residual_td_on_i  residuals_independency  residuals_normality  oos_mean  oos_mse  q  q_m2  m7  pct_outliers  frequency  arima_model
#> 
#> The variables exclusively found in the values matrix are:
#> frequency  arima_model
#> 
#> No score was calculated

# Extraire les modalités de la matrice
QR[["modalities"]]
#>    series residuals_homoskedasticity residuals_skewness residuals_kurtosis
#> 1  RF0610                       Good               Good               Good
#> 2  RF0620                        Bad                Bad                Bad
#> 3  RF0811                        Bad                Bad                Bad
#> 4  RF0812                       Good               Good               Good
#> 5  RF0893                       Good               Good               Good
#> 6  RF0899                       Good               Good               Good
#> 7  RF1011                       Good               Good               Good
#> 8  RF1012                       Good               Good               Good
#> 9  RF1013                        Bad                Bad                Bad
#> 10 RF1020                       Good               Good               Good
#> 11 RF1031                  Uncertain          Uncertain          Uncertain
#> 12 RF1032                       Good               Good               Good
#> 13 RF1039                  Uncertain          Uncertain          Uncertain
#>    qs_residual_sa_on_sa f_residual_sa_on_sa qs_residual_sa_on_i
#> 1                  Good                Good                Good
#> 2                  Good                Good                Good
#> 3                  Good                Good                Good
#> 4                  Good                Good                Good
#> 5                  Good                Good                Good
#> 6                  Good                Good                Good
#> 7                  Good              Severe                Good
#> 8                  Good              Severe                Good
#> 9                  Good                 Bad                Good
#> 10                 Good              Severe                Good
#> 11                 Good                 Bad                Good
#> 12                 Good                Good                Good
#> 13                 Good              Severe                Good
#>    f_residual_sa_on_i f_residual_td_on_sa f_residual_td_on_i
#> 1                Good           Uncertain               Good
#> 2                Good                Good               Good
#> 3                Good              Severe             Severe
#> 4                Good              Severe             Severe
#> 5                Good                Good               Good
#> 6                Good              Severe             Severe
#> 7              Severe              Severe             Severe
#> 8              Severe              Severe             Severe
#> 9                 Bad              Severe             Severe
#> 10                Bad              Severe             Severe
#> 11          Uncertain              Severe             Severe
#> 12               Good              Severe             Severe
#> 13                Bad              Severe                Bad
#>    residuals_independency residuals_normality  oos_mean   oos_mse    q q_m2
#> 1                    Good                Good      Good       Bad Good Good
#> 2                    Good                 Bad       Bad       Bad Good Good
#> 3                     Bad                 Bad      Good Uncertain  Bad  Bad
#> 4                     Bad                Good      Good      Good Good Good
#> 5               Uncertain                Good       Bad      Good  Bad  Bad
#> 6                     Bad                Good       Bad      Good Good Good
#> 7                     Bad                Good      Good      Good  Bad  Bad
#> 8                     Bad                Good       Bad      Good  Bad  Bad
#> 9                     Bad                 Bad       Bad      Good  Bad  Bad
#> 10                    Bad           Uncertain       Bad Uncertain Good  Bad
#> 11                    Bad                 Bad Uncertain       Bad Good Good
#> 12                    Bad                Good       Bad      Good  Bad  Bad
#> 13                    Bad                 Bad       Bad Uncertain Good Good
#>      m7 pct_outliers
#> 1  Good    Uncertain
#> 2  Good          Bad
#> 3  Good         Good
#> 4  Good         Good
#> 5  Good         Good
#> 6  Good         Good
#> 7  Good         Good
#> 8  Good         Good
#> 9   Bad         Good
#> 10 Good         Good
#> 11 Good    Uncertain
#> 12 Good         Good
#> 13 Good         Good
# Or:
QR[["modalities"]]
#>    series residuals_homoskedasticity residuals_skewness residuals_kurtosis
#> 1  RF0610                       Good               Good               Good
#> 2  RF0620                        Bad                Bad                Bad
#> 3  RF0811                        Bad                Bad                Bad
#> 4  RF0812                       Good               Good               Good
#> 5  RF0893                       Good               Good               Good
#> 6  RF0899                       Good               Good               Good
#> 7  RF1011                       Good               Good               Good
#> 8  RF1012                       Good               Good               Good
#> 9  RF1013                        Bad                Bad                Bad
#> 10 RF1020                       Good               Good               Good
#> 11 RF1031                  Uncertain          Uncertain          Uncertain
#> 12 RF1032                       Good               Good               Good
#> 13 RF1039                  Uncertain          Uncertain          Uncertain
#>    qs_residual_sa_on_sa f_residual_sa_on_sa qs_residual_sa_on_i
#> 1                  Good                Good                Good
#> 2                  Good                Good                Good
#> 3                  Good                Good                Good
#> 4                  Good                Good                Good
#> 5                  Good                Good                Good
#> 6                  Good                Good                Good
#> 7                  Good              Severe                Good
#> 8                  Good              Severe                Good
#> 9                  Good                 Bad                Good
#> 10                 Good              Severe                Good
#> 11                 Good                 Bad                Good
#> 12                 Good                Good                Good
#> 13                 Good              Severe                Good
#>    f_residual_sa_on_i f_residual_td_on_sa f_residual_td_on_i
#> 1                Good           Uncertain               Good
#> 2                Good                Good               Good
#> 3                Good              Severe             Severe
#> 4                Good              Severe             Severe
#> 5                Good                Good               Good
#> 6                Good              Severe             Severe
#> 7              Severe              Severe             Severe
#> 8              Severe              Severe             Severe
#> 9                 Bad              Severe             Severe
#> 10                Bad              Severe             Severe
#> 11          Uncertain              Severe             Severe
#> 12               Good              Severe             Severe
#> 13                Bad              Severe                Bad
#>    residuals_independency residuals_normality  oos_mean   oos_mse    q q_m2
#> 1                    Good                Good      Good       Bad Good Good
#> 2                    Good                 Bad       Bad       Bad Good Good
#> 3                     Bad                 Bad      Good Uncertain  Bad  Bad
#> 4                     Bad                Good      Good      Good Good Good
#> 5               Uncertain                Good       Bad      Good  Bad  Bad
#> 6                     Bad                Good       Bad      Good Good Good
#> 7                     Bad                Good      Good      Good  Bad  Bad
#> 8                     Bad                Good       Bad      Good  Bad  Bad
#> 9                     Bad                 Bad       Bad      Good  Bad  Bad
#> 10                    Bad           Uncertain       Bad Uncertain Good  Bad
#> 11                    Bad                 Bad Uncertain       Bad Good Good
#> 12                    Bad                Good       Bad      Good  Bad  Bad
#> 13                    Bad                 Bad       Bad Uncertain Good Good
#>      m7 pct_outliers
#> 1  Good    Uncertain
#> 2  Good          Bad
#> 3  Good         Good
#> 4  Good         Good
#> 5  Good         Good
#> 6  Good         Good
#> 7  Good         Good
#> 8  Good         Good
#> 9   Bad         Good
#> 10 Good         Good
#> 11 Good    Uncertain
#> 12 Good         Good
#> 13 Good         Good