Objets bilan qualité

QR_matrix permet de créer un objet de type QR_matrix contenant un bilan qualité.

QR_matrix creates a QR_matrix object containing a quality report.

QR_matrix(modalities = NULL, values = NULL, score_formula = NULL)

mQR_matrix(x = list(), ...)

is.QR_matrix(x)

is.mQR_matrix(x)

Arguments

modalities

a data.frame containing the output variables' modalities (Good, Bad, etc.)

values

a data.frame containing the output variables' values (test p-values, test statistics, etc.) Therefore, the values data frame can contain more variables than the data frame modalities.

score_formula

the formula used to calculate the series score (if defined).

x

a QR_matrix object, a mQR_matrix object or a list of QR_matrix objects.

...

objects of the same type as x.

Details

mQR_matrix permet de créer un objet de type mQR_matrix qui est une liste de bilans qualité (donc d'objets QR_matrix).

is.QR_matrix et is.mQR_matrix permettent de tester si un objet est un bilan qualité ou une liste de bilans qualité.

Un objet de type QR_matrix est une liste de trois paramètres :

le paramètre modalities est un data.frame contenant un ensemble de variables sous forme catégorielle (par défaut : Good, Uncertain, Bad, Severe).

le paramètre values est un data.frame contenant les valeurs associées aux indicateurs présents dans modalities (i.e. : p-valeurs, statistiques, etc.), ainsi que des variables qui n'ont pas de modalité (i.e. : fréquence de la série, modèle ARIMA, etc).

le paramètre score_formula contient la formule utilisée pour calculer le score (une fois le calcul réalisé).

mQR_matrix creates a mQR_matrix object containing a list of quality reports (ie. a list of QR_matrix objects).

is.QR_matrix and is.mQR_matrix are functions to test whether an object is a quality report or a list of quality reports.

AQR_matrix object is a list of three items:

modalities, a data.frame containing a set of categorical variables (by default: Good, Uncertain, Bad, Severe).

values, a data.frame containing the values corresponding to the modalities indicators (i.e. p-values, statistics, etc.), as well as variables for which a modality cannot be defined (e.g. the series frequency, the ARIMA model, etc).

score_formula contains the formula used to calculate the series score (once the calculus is done).

See also