Objets bilan qualité

QR_matrix permet de créer un objet de type QR_matrix contenant un bilan qualité.

QR_matrix(modalities = NULL, values = NULL, score_formula = NULL)

mQR_matrix(x = list(), ...)

is.QR_matrix(x)

is.mQR_matrix(x)

Arguments

modalities
un data.frame contenant les modalités (Good, Bad, etc.) associées aux variables.

values
un data.frame contenant les valeurs (p-valeurs des tests, statistiques, etc.) associées aux variables. Peut donc contenir plus de variables que le paramètre modalities.

score_formula
formule utilisée pour calculer le score global (s'il existe).

x
un objet de type QR_matrix, mQR_matrix ou une liste d'objets QR_matrix.

...
des objets du même type que x.

Details

mQR_matrix permet de créer un objet de type mQR_matrix qui est une liste de bilans qualité (donc d'objets QR_matrix).

is.QR_matrix et is.mQR_matrix permettent de tester si un objet est un bilan qualité ou une liste de bilans qualité.

Un objet de type QR_matrix est une liste de trois paramètres :

le paramètre modalities est un data.frame contenant un ensemble de variables sous forme catégorielle (par défaut : Good, Uncertain, Bad, Severe).

le paramètre values est un data.frame contenant les valeurs associées aux indicateurs présents dans modalities (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas de modalité (e.g. : fréquence de la série, modèle ARIMA, etc.).

le paramètre score_formula contient la formule utilisée pour calculer le score (une fois le calcul réalisé).