Exporting QR_matrix objects in an Excel file

To export a quality report in an Excel file.

# S3 method for QR_matrix
export_xlsx(
  x,
  layout = c("all", "modalities", "values", "combined"),
  create = TRUE,
  clear_sheet = TRUE,
  auto_format = TRUE,
  file_name,
  sheet_names,
  ...
)

Arguments

x

a QR_matrix object.

layout

the components of the report to export. By default, layout = "all": the matrices modalities ("modalities") and values ("values") are exported in separate files. To export them in a single file (in two sheets), use layout = "combined".

create

logical indicating whether to create an Excel file if it doesn't exist yet (create = TRUE by default)

clear_sheet

logical indicating whether to clear the Excel sheets before the export (clear_sheet = TRUE by default).

auto_format

logical indicating whether to format the output (auto_format = TRUE by default).

file_name

optional argument to choose the path and name of the file to export. If not specified, an export.xls will be created in the working directory.

sheet_names

names of the exported Excel sheets. If not specified, the sheets will be named after the exported components. If specified, existing sheets with these names will be overwritten.

...

other unused parameters.

See also