Exporting QR_matrix objects in an Excel file

To export a quality report in an Excel file.

Usage

# S3 method for class 'QR_matrix'
export_xlsx(x, file, auto_format = TRUE, overwrite = TRUE, ...)

Arguments

x

a QR_matrix object.

file

a character object with the path to the file to export que l'on veut créer

auto_format

logical indicating whether to format the output (auto_format = TRUE by default).

overwrite

logical indicating whether to create an Excel file if it doesn't exist yet (create = TRUE by default)

...

other unused arguments

Value

Returns invisibly (via invisible(x)) a workbook object created by XLConnect::loadWorkbook() for further manipulation.

See also