#' Extraction of a quality report

To extract a quality report from the csv file containing the diagnostics matrix.

extract_QR(matrix_output_file, sep = ";", dec = ",")

Arguments

matrix_output_file

the csv file containing the diagnostics matrix. When missing, a window opens to select a csv file.

sep

the separator used in the csv file (by default, sep = ";")

dec

the decimal separator used in the csv file (by default, dec = ",")

Value

a QR_matrix object.

Details

This function generates a quality report from a csv file containing diagnostics (usually from the file demetra_m.csv).

The demetra_m.csv file can be generated by launching the cruncher (functions cruncher or cruncher_and_param) with the default export parameters, having used the default option csv_layout = "vtable" to format the output tables of the functions cruncher_and_param and create_param_file when creating the parameters file.

This function returns a QR_matrix object, which is a list of 3 objects:

modalities, a data.frame containing several indicators and their categorical quality (Good, Uncertain, Bad, Severe).

values, a data.frame containing the same indicators and the values that lead to their quality category (i.e.: p-values, statistics, etc.) as well as additional variables that don't have a modality/quality (series frequency and arima model).

score_formula that will store the formula used to calculate the score (when relevant). Its initial value is NULL.

See also

Examples

if (FALSE) {
QR <- extract_QR()
print(QR)
# To extract the modalities matrix:
QR$modalities
# Or:
QR[["modalities"]]
}