Calcul d'un score pondéré pour chaque observation

Permet de pondérer un score déjà calculé en fonction des variables.

To weight a pre-calculated score

weighted_score(x, pond = 1)

Arguments

x

a QR_matrix or mQR_matrix object

pond

the weights to use. Can be an integer, a vector of integers, the name of one of the quality report variables or a list of weights for the mQR_matrix objects.

Value

L'objet en entrée avec le score recalculé

the input with additionnal weighted score

See also

Examples

if (FALSE) {
QR <- extract_QR()
QR <- compute_score(QR)
weighted_score(QR, 2) # Tous les scores sont multipliés par 2
}
if (FALSE) {
QR <- extract_QR()
QR <- compute_score(QR)
weighted_score(QR, 2) # All scores are multiplied by 2
}